refactor: transitionToMainContent()をPhotoTipsViewに移動した
